backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
google-generativeai==0.2.0
PyMuPDF==1.22.3
Werkzeug==2.3.2
python-dotenv==1.0.0  
gunicorn==21.2.0  
